AMD: Move MP1 to amd/common/mp and generalize it.
