+++ src/soc/amd/common/mp/src/lib.rs
@@ -1,0 +1,2 @@
+#![no_std]
+pub mod mpmailbox;
